gPTP: Fix correction applied to preciseOriginTimestamp
Change computation of correction applied to the preciseOriginTimestamp
by expressing correctionField and path delay in grandmaster time base
as per the IEEE 802.1AS-2011 standard (clause 11.1.3).

Fix bug where the correctionField and path delay are incorrectly
expressed in the local time base, which results in phase offsets that
oscillate in the range of tens of microseconds and never converge.