Reducing log noise when not asCapable
Reducing log level from error to debug for a few messages that are not actually
faulty to the protocol. They just result in the port not being asCapable. Also,
changing the behavior of a few log statements so that they only get printed
when the asCapable state changes. This avoids a lot of noise when gPTP is not
connected to an asCapable peer which helps users to identify what initially
caused gPTP to change states.