[gPTP] Restore port state check in processEvent
A recent change to the automotive code removed a port
state check that appeared to prevent forceAsCapable=false
from working correctly when an externalPortConfiguration
is set. Examining the code more closely, it appears that
the call to startPdelay was moved to syncDone() in cases
where gPTP's state is being restored from a file. Re-add
the port state check to restore that functionality, which
was originally added in 62e1451ee99310408d0e708f5039af53e4b052cd.