[gPTP] Always start PDelayReq and update IPC on INITIALIZE
Fixing a bug where gPTP wouldn't send PDelayReq messages when configured as a static slave. To resolve this, always start PDelayReq messages on INITIALIZE. The only times I can think of when you wouldn't want to do this would be if the link is down (see the pre-existing TODO in the code) or if you are reinitializing (in which case, PDelayReq messages should probably have been stopped first).

Adding logic to update the IPC on POWERUP/INITIALIZE. By doing this, clients can get the initial information from the daemon before any information changes or in case the information never changes (e.g., an externally configured slave with forceAsCapable).